requirements.txt
dbt-core
dbt-postgres
psycopg2-binary
python-dotenv
pandas
python-dotenv
